+++ src/lib.rs
@@ -4,7 +4,6 @@
 //! **pea2pea** is a simple, low-level, and customizable implementation of a TCP P2P node.
 
 mod config;
-mod known_peers;
 mod node;
 mod stats;
 mod topology;
@@ -13,8 +12,7 @@ pub mod connections;
 pub mod protocols;
 
 pub use config::Config;
-pub use connections::{Connection, ConnectionSide};
-pub use known_peers::KnownPeers;
+pub use connections::{Connection, ConnectionInfo, ConnectionSide};
 pub use node::Node;
 pub use stats::Stats;
 pub use topology::{connect_nodes, Topology};